fix React Router use, structure refactor (add langing page component), minor bug fixes
