+++ src/js/application/appMainPulpit.js
@@ -1,11 +1,11 @@
 import React, {useState} from 'react';
 import Logo from "../components/logo";
-import {BrowserRouter as Router, Switch, Route, Link} from 'react-router-dom'
+import {Route, Link} from 'react-router-dom'
 import Calendar from "./calendar";
 import Plans from "./plans";
 
 
-function MainPulpit() {
+function AppMainPulpit() {
   const [isActive, setActive] = useState(false)
   
   const toggleMenu = () => {
@@ -13,7 +13,6 @@ function MainPulpit() {
   }
   
   return (
-    <Router>
       <section className={'appScreen'}>
         <header className={'appHeader'}>
           <div className={'appHeader__logo'}>
@@ -24,11 +23,11 @@ function MainPulpit() {
               <div className={'appHeader__avatar'}/>
               <p className={'appHeader__userName'}>Username</p>
             </li>
-            <Link to={'/calendar'}><li>Calendar</li></Link>
-            <Link><li>History</li></Link>
-            <Link><li>Exercises</li></Link>
-            <Link to={'/plans'}><li>Plans</li></Link>
-            <Link><li>Log out</li></Link>
+            <Link  to={'/app/calendar'} onClick={toggleMenu} className={'link'}>Calendar</Link>
+            <Link onClick={toggleMenu} className={'link'}>History</Link>
+            <Link onClick={toggleMenu} className={'link'}>Exercises</Link>
+            <Link to={'/app/plans'} onClick={toggleMenu} className={'link'}>Plans</Link>
+            <Link onClick={toggleMenu} className={'link'}>Log out</Link>
           </ul>
           <div onClick={toggleMenu} className={'appHeader__burger'}>
             <div className={'line1'}/>
@@ -37,12 +36,11 @@ function MainPulpit() {
           </div>
         </header>
         <div className={'appPulpit'}>
-          <Route path={'/calendar'} component={Calendar}/>
-          <Route path={'/plans'} component={Plans}/>
+          <Route path={'/app/calendar'} component={Calendar}/>
+          <Route path={'/app/plans'} component={Plans}/>
         </div>
       </section>
-    </Router>
   );
 }
 
-export default MainPulpit;
+export default AppMainPulpit;
