+++ src/js/application/calendar.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 function Calendar() {
  return (
