+++ src/js/landing_page/logIn_section.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import CloseX from "../components/closeX";
+import {Link} from "react-router-dom";
 
 function LogInSection({isVisible, onVisibilityChange}) {
  
@@ -15,7 +16,7 @@ function LogInSection({isVisible, onVisibilityChange}) {
       <h2 className={'logIn__title'}>Personal Trainer!</h2>
       <input type={'text'} className={'logIn__input'} placeholder={'Username'}/>
       <input type={'text'} className={'logIn__input'} placeholder={'Password'}/>
-      <button className={'btn logIn__btn logIn__btn-log'}>LOG IN</button>
+      <Link to={'/app'} className={'btn logIn__btn logIn__btn-log'}>LOG IN</Link>
       <p className={'logIn__text'}>or</p>
       <button className={'btn logIn__btn logIn__btn-create'}>CREATE ACCOUNT</button>
     </div>
