+++ src/scss/application/_appMainPulpit.scss
@@ -51,13 +51,19 @@ appScreen{
       margin: 0;
     }
 
-    li {
+    .link {
+      text-align: center;
       padding: 20px 0;
+      font-family: $font-secondary;
+      font-size: 1.2rem;
+      color: $text-color_green;
+      width: 100%;
+
       @include tablet(660px) {
         padding: 0 0 0 20px;
       }
       @include desktop{
-        padding: 15px 0;
+        padding: 15px 0 0 20px;
       }
     }
 
@@ -121,7 +127,11 @@ appScreen{
   position: absolute;
   bottom: 0;
   right: 0;
-  width: calc(100% - 150px);
+  width: 100%;
   height: calc(100vh - 100px);
   border: none;
+  z-index: -2;
+  @include desktop{
+    width: calc(100% - 150px);
+  }
 }
