+++ src/scss/elements/_button.scss
@@ -5,6 +5,7 @@
   padding: 10px 15px;
   color: $text-color_light;
   font-family: $font-primary;
+text-align: center;
   @include tablet{
     margin-top: 30px;
     padding: 20px 25px ;
