+++ src/scss/main.scss
@@ -16,4 +16,4 @@
 @import "landing_page/logIn_section";
 
 //application
-@import "application/mainPulpit";
+@import "application/appMainPulpit";
